model-ai/requirements.txt
sentence-transformers
fastapi
uvicorn
load_dotenv
pydantic_settings
chromadb